requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Werkzeug==3.0.1
requests==2.31.0
psutil==5.9.6
python-dotenv==1.0.0
bleach==6.1.0

